+++ src/app/proposals/shared/types/index.ts
@@ -94,3 +94,36 @@ export enum MilestoneLabels {
   THIRD = 'Milestone 3',
   NO_MILESTONE = 'No milestone',
 }
+
+export interface Delegator {
+  id: string
+}
+
+interface VoteCast {
+  id: string
+}
+
+interface Account {
+  delegatedVotes: string
+  delegators: Delegator[]
+  VoteCasts: VoteCast[]
+}
+
+export interface Contributor {
+  id: string
+  account: Account
+  createdAt: string
+  nftId: string
+}
+
+export interface ContributorGraphResponse {
+  contributors: Contributor[]
+}
+
+export interface ContributorApiResponse {
+  id: string
+  delegatedVotes: string
+  delegators: number // flattened ids
+  votes: number
+  createdAt: string
+}
